+++ src/resources/assets/js/jarboe.js
@@ -82,6 +82,45 @@ const Jarboe = {
         });
     },
 
+    bigToastSuccess: function(title, content, timeout) {
+        this.bigToast({
+            title: title,
+            content: content,
+            timeout: timeout,
+            color: '#739E73',
+            icon: 'fa fa-check',
+        });
+    },
+
+    bigToastDanger: function(title, content, timeout) {
+        this.bigToast({
+            title: title,
+            content: content,
+            timeout: timeout,
+            color: '#C46A69',
+            icon: 'fa fa-warning shake animated',
+        });
+    },
+
+    bigToastWarning: function(title, content, timeout) {
+        this.bigToast({
+            title: title,
+            content: content,
+            timeout: timeout,
+            color: '#C79121',
+            icon: 'fa fa-shield fadeInLeft animated',
+        });
+    },
+
+    bigToastInfo: function(title, content, timeout) {
+        this.bigToast({
+            title: title,
+            content: content,
+            timeout: timeout,
+            color: '#3276B1',
+            icon: 'fa fa-bell swing animated',
+        });
+    },
 
     /**
      * Object that hold functions for initialising plugins and stuff for fields.